[유저] 모바일 로그인 UI 변경 (#824)
