+++ src/pages/Auth/AuthPage/Auth.module.scss
@@ -13,38 +13,6 @@
 
   @include media.media-breakpoint(mobile) {
     text-align: center;
-    display: flex;
-    flex-direction: column;
-    position: relative;
-  }
-
-  &__header {
-    @include media.media-breakpoint(mobile) {
-      position: absolute;
-      top: 16px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      width: 100%;
-    }
-  }
-
-  &__title {
-    @include media.media-breakpoint(mobile) {
-      font-size: 1.125rem;
-      font-weight: 500;
-      line-height: 1.6;
-      height: fit-content;
-    }
-  }
-
-  &__button {
-    @include media.media-breakpoint(mobile) {
-      position: absolute;
-      left: 0;
-      display: flex;
-      margin-left: 16px;
-    }
   }
 
   &__content {
@@ -56,10 +24,7 @@
     justify-content: center;
 
     @include media.media-breakpoint(mobile) {
-      width: 100%;
-      margin: 0;
-      min-width: 360px;
-      height: 100vh;
+      margin: 0 48px;
     }
   }
 
@@ -68,8 +33,8 @@
     margin-bottom: 40px;
 
     @include media.media-breakpoint(mobile) {
-      margin-bottom: 60px;
-      margin-right: 17px;
+      margin-bottom: 32px;
+      align-self: flex-start;
     }
   }
 
@@ -79,31 +44,6 @@
 
     @include media.media-breakpoint(mobile) {
       width: 105px;
-    }
-  }
-
-  &__go-back {
-    position: absolute;
-    top: 20px;
-    left: 15px;
-    background-color: transparent;
-    border: none;
-    padding: 0;
-    cursor: pointer;
-
-    @include media.media-breakpoint(mobile) {
-      top: 15px;
-      left: 10px;
-    }
-  }
-
-  &__left-arrow-image {
-    width: 35px;
-    height: 35px;
-
-    @include media.media-breakpoint(mobile) {
-      width: 30px;
-      height: 30px;
     }
   }
 }
