+++ src/pages/Auth/AuthPage/index.tsx
@@ -1,10 +1,17 @@
-import { Outlet } from 'react-router-dom';
+import { Outlet, Link, useLocation } from 'react-router-dom';
+import ROUTES from 'static/routes';
 import styles from './Auth.module.scss';
 
 function AuthPage() {
+  const location = useLocation();
   return (
     <div className={styles.template}>
       <div className={styles.template__content}>
+        {location.pathname === ROUTES.Auth() && (
+        <Link className={styles.template__logo} to={ROUTES.Main()}>
+          <img className={styles.template__image} src="https://static.koreatech.in/assets/img/logo_primary.png" alt="main logo" />
+        </Link>
+        )}
         <Outlet />
       </div>
     </div>
