+++ src/pages/Auth/LoginPage/LoginPage.module.scss
@@ -1,170 +1,7 @@
-@use "src/utils/scss/media" as media;
-
 .template {
   display: flex;
   justify-content: center;
   align-items: center;
   flex-direction: column;
-
-  &__copyright {
-    font-family: NanumBarunGothic, sans-serif;
-    margin-top: 28px;
-    color: gray;
-    letter-spacing: -0.7px;
-    font-size: 13px;
-    text-align: center;
-
-    @include media.media-breakpoint(mobile) {
-      margin-top: 8px;
-      width: auto;
-      font-family: NanumBarunGothic, sans-serif;
-      font-size: 11px;
-      font-weight: normal;
-      color: #7d7d7d;
-    }
-  }
-}
-
-.loginform {
-  width: auto;
-
-  &__button {
-    width: 390px;
-    height: 45px;
-    border-radius: 0;
-    line-height: 1.3;
-    background-color: #175c8e;
-    color: #fff;
-    font-family: NanumBarunGothic, sans-serif;
-    font-size: 20px;
-    letter-spacing: -1px;
-    font-weight: 700;
-    cursor: pointer;
-    border: 1px solid #175c8e;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    &:disabled {
-      background: #e5eaf0;
-      border: solid 1px #d2dae2;
-    }
-
-    @include media.media-breakpoint(mobile) {
-      width: 320px;
-      padding: 0;
-      margin: 0 auto;
-      height: 36px;
-      font-size: 15px;
-    }
-  }
-}
-
-.form-input {
-  width: 390px;
-  height: 45px;
-  box-sizing: border-box;
-  border: #d2dae2 1px solid;
-  font-family: NanumBarunGothic, sans-serif;
-  font-size: 15px;
-  padding-left: 20px;
-  margin-bottom: 12px;
-
-  &::placeholder {
-    font-size: 15px;
-    letter-spacing: -0.8px;
-    color: #bec9d5;
-  }
-
-  &:disabled {
-    background-color: #e5eaf0;
-    border: 1px solid #d2dae2;
-    color: #b5c1cd;
-  }
-
-  @include media.media-breakpoint(mobile) {
-    width: 320px;
-    height: 36px;
-    padding-left: 16px;
-
-    &::placeholder {
-      font-size: 14px;
-    }
-  }
-}
-
-.auto-login {
   width: 100%;
-  height: 43px;
-  text-align: left;
-
-  @include media.media-breakpoint(mobile) {
-    width: 320px;
-    height: 52px;
-    margin: 0 auto;
-  }
-
-  &__label {
-    margin-top: 10px;
-    font-size: 12px;
-    color: #858585;
-    letter-spacing: -0.6px;
-    cursor: pointer;
-    display: inline-block;
-    position: relative;
-    line-height: 16px;
-    user-select: none;
-  }
-
-  @include media.media-breakpoint(mobile) {
-    margin-top: 18px;
-  }
-
-  &__checkbox {
-    cursor: pointer;
-    width: 14px;
-    height: 14px;
-    margin: 0 8px 0 0;
-    position: relative;
-    top: 3px;
-    background-color: #fff;
-    border: 1px solid #d2dae2;
-    border-radius: 0;
-    appearance: none;
-    box-sizing: content-box;
-  }
-
-  &__checkbox:checked {
-    content: "";
-    background-image: url("https://static.koreatech.in/assets/img/check.png");
-    background-size: cover;
-  }
-}
-
-.help {
-  border-top: 1px #d2dae2 solid;
-  width: 100%;
-  height: 43px;
-  padding-top: 18px;
-  font-size: 12px;
-  letter-spacing: -0.6px;
-  color: #858585;
-  text-align: center;
-
-  @include media.media-breakpoint(mobile) {
-    width: 320px;
-    margin: 0 auto;
-  }
-
-  &__link {
-    text-decoration: none;
-    color: #858585;
-    padding-right: 19px;
-    padding-left: 14px;
-    cursor: pointer;
-
-    & + & {
-      border-left: 1px #d2dae2 solid;
-    }
-  }
 }
