+++ src/pages/Auth/LoginPage/components/AdditionalLink/AdditionalLink.module.scss
@@ -1,0 +1,46 @@
+@use "src/utils/scss/media" as media;
+
+.help {
+  border-top: 1px #d2dae2 solid;
+  width: 100%;
+  height: 43px;
+  padding-top: 18px;
+  font-size: 12px;
+  letter-spacing: -0.6px;
+  color: #858585;
+  text-align: center;
+
+  @include media.media-breakpoint(mobile) {
+    display: flex;
+    justify-content: center;
+    width: 320px;
+    height: 32px;
+    margin-top: 32px;
+    padding: 0;
+    border: 0;
+  }
+
+  &__link {
+    text-decoration: none;
+    color: #858585;
+    padding-right: 19px;
+    padding-left: 14px;
+    cursor: pointer;
+
+    @include media.media-breakpoint(mobile) {
+      font-size: 13px;
+      font-weight: 400;
+      color: #a1a1a1;
+      display: flex;
+      gap: 4px;
+      justify-content: center;
+      align-items: center;
+      height: 20px;
+      padding: 0 8px;
+    }
+
+    & + & {
+      border-left: 1px #d2dae2 solid;
+    }
+  }
+}
