+++ src/pages/Auth/LoginPage/components/LoginForm/LoginForm.module.scss
@@ -1,0 +1,148 @@
+@use "src/utils/scss/media" as media;
+
+.loginform {
+  width: 100%;
+
+  &__button {
+    width: 390px;
+    height: 45px;
+    line-height: 1.3;
+    background-color: #175c8e;
+    color: #fff;
+    font-family: NanumBarunGothic, sans-serif;
+    font-size: 20px;
+    letter-spacing: -1px;
+    font-weight: 700;
+    cursor: pointer;
+    border: 1px solid #175c8e;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    &:disabled {
+      background: #e5eaf0;
+      border: solid 1px #d2dae2;
+    }
+
+    @include media.media-breakpoint(mobile) {
+      width: 100%;
+      padding: 0;
+      margin: 24px auto 0;
+      height: 44px;
+      border-radius: 8px;
+      background-color: #175c8e;
+      border: 0;
+      font-size: 15px;
+      font-weight: 500;
+    }
+  }
+
+  &__first-button {
+    @include media.media-breakpoint(mobile) {
+      margin: 36px auto 0;
+      background-color: #f7941e;
+    }
+  }
+}
+
+.form-input-wrapper {
+  position: relative;
+  height: 45px;
+  margin-bottom: 12px;
+
+  @include media.media-breakpoint(mobile) {
+    height: 40px;
+  }
+}
+
+.form-input {
+  width: 390px;
+  height: 45px;
+  box-sizing: border-box;
+  border: #d2dae2 1px solid;
+  font-family: NanumBarunGothic, sans-serif;
+  font-size: 15px;
+  padding-left: 20px;
+  margin-bottom: 12px;
+
+  &::placeholder {
+    font-size: 15px;
+    letter-spacing: -0.8px;
+    color: #bec9d5;
+  }
+
+  &:disabled {
+    background-color: #e5eaf0;
+    border: 1px solid #d2dae2;
+    color: #b5c1cd;
+  }
+
+  @include media.media-breakpoint(mobile) {
+    width: 100%;
+    height: 40px;
+    padding-left: 16px;
+    border-width: 0 0 1px;
+
+    &::placeholder {
+      font-size: 14px;
+    }
+  }
+}
+
+.password-toggle-button {
+  position: absolute;
+  max-height: 20px;
+  max-width: 20px;
+  top: 50%;
+  right: 10px;
+  transform: translateY(-50%);
+}
+
+.auto-login {
+  width: 100%;
+  height: 43px;
+  text-align: left;
+
+  @include media.media-breakpoint(mobile) {
+    width: 100%;
+    height: 100%;
+    text-align: right;
+  }
+
+  &__label {
+    margin-top: 10px;
+    font-size: 12px;
+    color: #858585;
+    letter-spacing: -0.6px;
+    cursor: pointer;
+    display: inline-block;
+    position: relative;
+    line-height: 16px;
+    user-select: none;
+
+    @include media.media-breakpoint(mobile) {
+      margin-top: 24px;
+    }
+  }
+
+  &__checkbox {
+    cursor: pointer;
+    width: 14px;
+    height: 14px;
+    margin: 0 8px 0 0;
+    position: relative;
+    top: 3px;
+    background-color: #fff;
+    border: 1px solid #d2dae2;
+    border-radius: 0;
+    appearance: none;
+    box-sizing: content-box;
+  }
+
+  &__checkbox:checked {
+    content: "";
+    background-image: url("https://static.koreatech.in/assets/img/check.png");
+    background-size: cover;
+    margin: 0;
+  }
+}
