+++ src/pages/Auth/SignupPage/hooks/useStep.tsx
@@ -1,9 +1,11 @@
 /* eslint-disable react/jsx-no-useless-fragment */
 import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 function useStep<T extends string>(initialStep: T) {
   const [currentStep, setCurrentStep] = useState<T>(initialStep);
   const [historyStep, setHistoryStep] = useState<T[]>([initialStep]);
+  const navigate = useNavigate();
 
   const nextStep = (next: T) => {
     setCurrentStep(next);
@@ -11,7 +13,10 @@ function useStep<T extends string>(initialStep: T) {
   };
 
   const goBack = () => {
-    if (historyStep.length <= 1) return;
+    if (historyStep.length <= 1) {
+      navigate(-1);
+      return;
+    }
     setHistoryStep((prev) => {
       const newHistory = prev.slice(1);
       setCurrentStep(newHistory[0]);
