+++ src/static/auth.ts
@@ -1,6 +1,7 @@
 export const REGEX = {
   PASSWORD: /^(?=.*[A-Za-z])(?=.*\d)(?=.*[!@#$%^&*()_\-+=~`[\]{}|\\:;"'<>,.?/])[A-Za-z\d!@#$%^&*()_\-+=~`[\]{}|\\:;"'<>,.?/]{6,18}$/,
   NICKNAME: /^[가-힣a-zA-Z0-9]{1,10}$/,
+  EMAIL: /^[a-z_0-9]{1,12}$/,
 };
 
 export const MESSAGES = {
